math-servise/src/test/java/com/skodin/services/QuestionServiceTest.java
package com.skodin.services;

import com.skodin.models.Question;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.List;
import java.util.Random;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

@ExtendWith(MockitoExtension.class)
class QuestionServiceTest {

    @InjectMocks
    private QuestionService questionService;

    @Test
    void getQuestions_ReturnsListOfQuestions() {
        int count = 5;

        List<Question> questions = questionService.getQuestions(count);

        assertNotNull(questions);
        assertEquals(count, questions.size());
    }
}
